requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
tavily-python>=0.5.0
langgraph>=0.2.0
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
# Vector Database
chromadb>=0.4.24
openai>=1.0.0

# HuggingFace Embeddings
sentence-transformers>=2.5.1
